chatbot/requirements.txt
pypdf
langchain
langchain-community>=0.3.7
langchain-chroma
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
sentence-transformers
timedelta
langchain_huggingface
langchain-ollama
langgraph
pdfplumber
whoosh
elasticsearch==8.11.1
elasticsearch-dsl==8.15.4
django-elasticsearch-dsl==8.0
django-rest-passwordreset
